+++ packages/newspringchurchapp/src/prayer/PrayerTab/index.js
@@ -1,0 +1,3 @@
+import PrayerTab from './PrayerTab';
+
+export default PrayerTab;
